refactor(agent): extract execution supervisor type definitions
